enable multi-label with type double

+++ include/caffe/util/io.hpp
@@ -85,38 +85,38 @@ inline void WriteProtoToBinaryFile(
   WriteProtoToBinaryFile(proto, filename.c_str());
 }
 
-bool ReadFileToDatum(const string& filename, const std::vector<int> label, Datum* datum);
+bool ReadFileToDatum(const string& filename, const std::vector<double> label, Datum* datum);
 
 inline bool ReadFileToDatum(const string& filename, Datum* datum) {
-  return ReadFileToDatum(filename, std::vector<int>(1,-1), datum);
+  return ReadFileToDatum(filename, std::vector<double>(1,-1), datum);
 }
 
-bool ReadImageToDatum(const string& filename, const std::vector<int> label,
+bool ReadImageToDatum(const string& filename, const std::vector<double> label,
     const int height, const int width, const bool is_color,
     const std::string & encoding, Datum* datum);
 
-inline bool ReadImageToDatum(const string& filename, const std::vector<int> label,
+inline bool ReadImageToDatum(const string& filename, const std::vector<double> label,
     const int height, const int width, const bool is_color, Datum* datum) {
   return ReadImageToDatum(filename, label, height, width, is_color,
                           "", datum);
 }
 
-inline bool ReadImageToDatum(const string& filename, const std::vector<int> label,
+inline bool ReadImageToDatum(const string& filename, const std::vector<double> label,
     const int height, const int width, Datum* datum) {
   return ReadImageToDatum(filename, label, height, width, true, datum);
 }
 
-inline bool ReadImageToDatum(const string& filename, const std::vector<int> label,
+inline bool ReadImageToDatum(const string& filename, const std::vector<double> label,
     const bool is_color, Datum* datum) {
   return ReadImageToDatum(filename, label, 0, 0, is_color, datum);
 }
 
-inline bool ReadImageToDatum(const string& filename, const std::vector<int> label,
+inline bool ReadImageToDatum(const string& filename, const std::vector<double> label,
     Datum* datum) {
   return ReadImageToDatum(filename, label, 0, 0, true, datum);
 }
 
-inline bool ReadImageToDatum(const string& filename, const std::vector<int> label,
+inline bool ReadImageToDatum(const string& filename, const std::vector<double> label,
     const std::string & encoding, Datum* datum) {
   return ReadImageToDatum(filename, label, 0, 0, true, encoding, datum);
 }
