+++ src/caffe/util/io.cpp
@@ -111,7 +111,7 @@ static bool matchExt(const std::string & fn,
     return true;
   return false;
 }
-bool ReadImageToDatum(const string& filename, const std::vector<int> label,
+bool ReadImageToDatum(const string& filename, const std::vector<double> label,
     const int height, const int width, const bool is_color,
     const std::string & encoding, Datum* datum) {
   cv::Mat cv_img = ReadImageToCVMat(filename, height, width, is_color);
@@ -144,7 +144,7 @@ bool ReadImageToDatum(const string& filename, const std::vector<int> label,
   }
 }
 
-bool ReadFileToDatum(const string& filename, const std::vector<int> label,
+bool ReadFileToDatum(const string& filename, const std::vector<double> label,
     Datum* datum) {
   std::streampos size;
